[CELEBORN-1549] Fix networkLocation persistence into Ratis
### What changes were proposed in this pull request?
Fixing a bug where the `networkLocation` is not persisted in Ratis, and the master defaults to `DEFAULT_RACK` when it loads the snapshot. This was missed in https://github.com/apache/celeborn/pull/2367 unfortunately, and it came up during our stress testing internally.

### Why are the changes needed?
Needed for custom network aware replication, so that networkLocation state is kept in snapshot file.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Updated unit test to ensure serde is correct.

Closes #2669 from akpatnam25/CELEBORN-1549.

Authored-by: Aravind Patnam <apatnam@linkedin.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>