[CELEBORN-1553] Using the request base url as swagger server
### What changes were proposed in this pull request?

Using the request base url as swagger server, to prevent the swagger server not reachable and `CORS` error if the swagger server urls do not match.

Currently, if the http host is bound to local, the swagger server is not reachable.
For example:
```
celeborn.master.http.host=0.0.0.0
celeborn.worker.http.host=0.0.0.0
```

### Why are the changes needed?
![image](https://github.com/user-attachments/assets/3553fb94-c937-4877-bd55-3df0c121bd01)

### Does this PR introduce _any_ user-facing change?

No, just use the request base url as swagger server.

### How was this patch tested?
Integration testing:
<img width="1483" alt="image" src="https://github.com/user-attachments/assets/f8465bcb-a266-4532-9f11-52e9374c56a5">

<img width="1423" alt="image" src="https://github.com/user-attachments/assets/84eacd86-7ba3-4f14-907b-6d529c6e0e28">

Closes #2674 from turboFei/swagger_server.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>